[CIAS30-3822] Extract FormikTimezoneSelect.tsx component
