+++ app/components/FormikTimezoneSelect/utils.ts
@@ -1,0 +1,4 @@
+import dayjs from 'dayjs';
+
+export const getTimezoneLabel = (timeZone: string) =>
+  `${timeZone} (UTC ${dayjs().tz(timeZone).format('Z')})`;
