+++ app/components/TimezoneForm/TimezoneForm.js
@@ -1,23 +1,18 @@
-import React, { useMemo } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { Formik } from 'formik';
-import dayjs from 'dayjs';
 
-import { TIMEZONES } from 'utils/timezones';
-
-import FormikSelect from 'components/FormikSelect';
-
-import messages from './messages';
-
-const timezoneLabel = (timeZone) =>
-  `${timeZone} (UTC ${dayjs().tz(timeZone).format('Z')})`;
+import {
+  FormikTimezoneSelect,
+  getTimezoneLabel,
+} from 'components/FormikTimezoneSelect';
 
 const initialValues = (timeZone) => {
   if (!timeZone) return { timeZone: null };
   return {
     timeZone: {
       value: timeZone,
-      label: timezoneLabel(timeZone),
+      label: getTimezoneLabel(timeZone),
     },
   };
 };
@@ -35,15 +30,6 @@ export const TimezoneForm = ({
     setSubmitting(false);
   };
 
-  const timezoneOptions = useMemo(
-    () =>
-      TIMEZONES.map((elem) => ({
-        value: elem,
-        label: timezoneLabel(elem),
-      })),
-    [TIMEZONES],
-  );
-
   return (
     <Formik
       initialValues={initialValues(timeZone)}
@@ -51,11 +37,9 @@ export const TimezoneForm = ({
       enableReinitialize
     >
       {() => (
-        <FormikSelect
+        <FormikTimezoneSelect
           columnStyleProps={props}
           formikKey="timeZone"
-          label={formatMessage(messages.timeZoneLabel)}
-          options={timezoneOptions}
           disabled={disabled}
           submitOnChange
         />
