+++ app/translations/ar.json
@@ -200,6 +200,7 @@
   "app.components.FormikPhoneNumberInput.phoneNumberLabel": "",
   "app.components.FormikPhoneNumberInput.phoneNumberPrefixLabel": "",
   "app.components.FormikPhoneNumberInput.phoneNumberRequired": "",
+  "app.components.FormikTimezoneSelect.timeZoneLabel": "",
   "app.components.HenryFordBranchingInfoModal.content": "",
   "app.components.HenryFordBranchingInfoModal.description": "",
   "app.components.HenryFordBranchingInfoModal.okay": "",
@@ -1590,7 +1591,6 @@
   "app.containers.ThirdParty.reportsToSend": "",
   "app.containers.ThirdParty.zeroReportChosen": "",
   "app.containers.TileMapper.noResults": "",
-  "app.containers.TimezoneForm.timeZoneLabel": "",
   "app.containers.TlfbConfig.dateRange": "",
   "app.containers.TlfbConfig.displayHelpingMaterials": "",
   "app.containers.TlfbConfig.noOfDaysLabel": "",
