+++ app/translations/en.json
@@ -200,6 +200,7 @@
   "app.components.FormikPhoneNumberInput.phoneNumberLabel": "Phone number",
   "app.components.FormikPhoneNumberInput.phoneNumberPrefixLabel": "Phone prefix",
   "app.components.FormikPhoneNumberInput.phoneNumberRequired": "Number is required",
+  "app.components.FormikTimezoneSelect.timeZoneLabel": "Timezone",
   "app.components.HenryFordBranchingInfoModal.content": "It looks like you want to duplicate {type, select, session {a session in the intervention} other {the intervention}} which includes Henry Ford's integration. <bold>Access to the Henry Ford's integration for the copy will need to be granted by a CIAS admin again</bold>. Also, Henry Ford Initial Screens will be removed from the copy together with all of their branching connections.",
   "app.components.HenryFordBranchingInfoModal.description": "Henry Ford Branching",
   "app.components.HenryFordBranchingInfoModal.okay": "Okay, I understand",
@@ -1590,7 +1591,6 @@
   "app.containers.ThirdParty.reportsToSend": "Reports to send",
   "app.containers.ThirdParty.zeroReportChosen": "No reports were chosen for that answer",
   "app.containers.TileMapper.noResults": "There are no reports from your sessions",
-  "app.containers.TimezoneForm.timeZoneLabel": "Timezone",
   "app.containers.TlfbConfig.dateRange": "Date range",
   "app.containers.TlfbConfig.displayHelpingMaterials": "Display standard drink chart to the participants",
   "app.containers.TlfbConfig.noOfDaysLabel": "Type the number of days",
